scripts/news/requirements.txt
feedparser>=6.0.11
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
